+++ CHANGELOG.md
@@ -6,6 +6,8 @@ CHANGELOG
 * Add constants for EC2 Seurity Group Protocols
 * Add constants for ALB IpAddressType and LoadBalancer type
 * Add constants for Route53 record types
+* Subscription resources will now be parented by default by the resource they were created off of.
+  This has been implemented using 'aliases' so this will not have any effect on existing stacks.
 
 ---